New performance improvement for OutputTree
As the output trees of non-observable FSMs can contain the same output
sequence multiple times, the contains relation needed to be modified
compared to the first version of this commit trying to improve the
performance of the contains check:
The std::includes() function does not handle sets well where equal
elements can be found in higher amounts in the sorted range which is to
be included than in the sorted range to include the other range.
Now, a call to std::unique() removes all non-unique elements of both
ranges so that only unique output traces are in either set.  Afterwards,
the call to std::includes() behaves like real set inclusion.